Updated heirachy, table methods, example
The object heirachy has been modified heavily.

Table methods can now return a JSON object with a docs parameter instead
of an array of ids. docs should be the array of ids instead. The object
will be passed as the second argument to the success callback.

The example has been heavily modified to support the new users
functionality. Updates to the group bits will follow